requirements.txt
anthropic==0.28.0
streamlit==1.35.0
python-dotenv==1.0.1